Add line numbers to termbox frontend
This is a first pass implementation at adding line number rendering to
the termbox frontend. There is some refactoring I would like to do to
that entire part of the render loop as well as the line number
rendering,
but I feel that would be best saved for another commit.

This PR addresses issue #164