Add gofmt script for testing
If Go source code is checked into the repository but not properly
formatted using gofmt, this new script will print the aberrant files and
exit 1 to fail Travis tests.

It's listed as a new script in .travis.yml. All scripts are run (even if
any of them exit nonzero), so regular tests still occur if the gofmt
check fails.

The script also ignores the 3rdparty directory, since that contains git
submodules (some of which have unformatted Go code in them). Since those
repositories are not under our control, we don't want to check them for
formatting.